+++ x-uploader/test/x/uploader/media_chunked_test.rb
@@ -39,34 +39,29 @@ module X
     end
 
     def test_append_uploads_chunks_with_segment_indices
-      chunk_paths = Uploader::Media.send(:split, VIDEO_FILE, @video_size - 1)
       stub_append_request
 
-      Uploader::Media.send(:append, client: @client, file_paths: chunk_paths, media: media_hash, boundary: TEST_BOUNDARY)
+      append(chunk_size: @video_size - 1)
       bodies = collect_request_bodies(:post, append_url, expected_count: 2)
 
-      chunk_paths.each_index { |i| assert_segment_in_bodies(bodies, i) }
+      2.times { |i| assert_segment_in_bodies(bodies, i) }
     end
 
     def test_append_does_not_report_exceptions_on_stderr
-      chunk_paths = Uploader::Media.send(:split, VIDEO_FILE, @video_size)
       stub_request(:post, append_url).to_return(status: 401)
 
       _, err = capture_subprocess_io do
-        assert_raises(Unauthorized) do
-          Uploader::Media.send(:append, client: @client, file_paths: chunk_paths, media: media_hash, boundary: TEST_BOUNDARY)
-        end
+        assert_raises(Unauthorized) { append(chunk_size: @video_size) }
       end
 
       assert_empty err
     end
 
     def test_append_leaves_global_exception_reporting_alone
-      chunk_paths = Uploader::Media.send(:split, VIDEO_FILE, @video_size)
       stub_append_request
       previous = Thread.report_on_exception
       Thread.report_on_exception = true
-      Uploader::Media.send(:append, client: @client, file_paths: chunk_paths, media: media_hash, boundary: TEST_BOUNDARY)
+      append(chunk_size: @video_size)
 
       assert Thread.report_on_exception
     ensure
@@ -74,6 +69,10 @@ module X
     end
 
     private
+
+    def append(chunk_size:, **)
+      Uploader::Media.send(:append, client: @client, file_path: VIDEO_FILE, chunk_size:, media: media_hash, boundary: TEST_BOUNDARY, **)
+    end
 
     def media_hash = {"id" => TEST_MEDIA_ID}
     def json_headers = {"content-type" => "application/json"}